CR1 - move function to resourcescaler
